+++ packages/stream_video/lib/src/webrtc/model/rtc_tracks_info.dart
@@ -11,6 +11,9 @@ class RtcTrackInfo {
     required this.codec,
     required this.muted,
     required this.publishOptionId,
+    required this.dtx,
+    required this.stereo,
+    required this.red,
   });
 
   final String? trackId;
@@ -20,11 +23,19 @@ class RtcTrackInfo {
   final SfuCodec? codec;
   final bool muted;
   final int publishOptionId;
+  final bool stereo;
+
+  /// When true, the call will optimize bandwidth usage by disabling audio when participants are not speaking via discontinuous transmission (DTX).
+  final bool dtx;
+
+  /// When true, the call will send additional audio streams to ensure calls are more resistant to poor network conditions.
+  final bool red;
 
   @override
   String toString() {
     return 'RtcTrackInfo{trackId: $trackId, trackType: $trackType, '
-        'mid: $mid, layers: $layers, codec: $codec, muted: $muted, publishOptionId: $publishOptionId}';
+        'mid: $mid, layers: $layers, codec: $codec, muted: $muted, publishOptionId: $publishOptionId, '
+        'dtx: $dtx, stereo: $stereo, red: $red}';
   }
 }